--- global.json
@@ -1,6 +1,0 @@
-{
-  "sdk": {
-    "version": "9.0.203",
-    "rollForward": "latestFeature"
-  }
-}
