+++ src/Stack/Infrastructure/ConsoleOutputProvider.cs
@@ -1,0 +1,9 @@
+namespace Stack.Infrastructure;
+
+public class ConsoleOutputProvider : IOutputProvider
+{
+    public async Task WriteLine(string output, CancellationToken cancellationToken = default)
+    {
+        await Console.Out.WriteLineAsync(output.AsMemory(), cancellationToken);
+    }
+}
